+++ tests/apis-changes.js
@@ -8,7 +8,7 @@ const suite = getSuite({ name: 'Testing Observable APIs - Changes' });
 suite.runTest({ name: 'verify object - root - insert' }, test => {
 	let c;
 	const o = Observable.from({});
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o.some = 'new';
 	assertObjectEqual(test, c, { type: 'insert', path: ['some'], value: 'new', oldValue: undefined, object: o });
 });
@@ -16,7 +16,7 @@ suite.runTest({ name: 'verify object - root - insert' }, test => {
 suite.runTest({ name: 'verify object - deep - insert' }, test => {
 	let c;
 	const o = Observable.from({ a: {} });
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o.a.some = 'new';
 	assertObjectEqual(test, c, { type: 'insert', path: ['a', 'some'], value: 'new', oldValue: undefined, object: o.a });
 });
@@ -24,7 +24,7 @@ suite.runTest({ name: 'verify object - deep - insert' }, test => {
 suite.runTest({ name: 'verify object - root - update' }, test => {
 	let c;
 	const o = Observable.from({ p: 'old' });
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o.p = 'new';
 	assertObjectEqual(test, c, { type: 'update', path: ['p'], value: 'new', oldValue: 'old', object: o });
 });
@@ -32,7 +32,7 @@ suite.runTest({ name: 'verify object - root - update' }, test => {
 suite.runTest({ name: 'verify object - deep - update' }, test => {
 	let c;
 	const o = Observable.from({ a: { p: 'old' } });
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o.a.p = 'new';
 	assertObjectEqual(test, c, { type: 'update', path: ['a', 'p'], value: 'new', oldValue: 'old', object: o.a });
 });
@@ -40,7 +40,7 @@ suite.runTest({ name: 'verify object - deep - update' }, test => {
 suite.runTest({ name: 'verify object - root - delete' }, test => {
 	let c;
 	const o = Observable.from({ p: 'old' });
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	delete o.p;
 	assertObjectEqual(test, c, { type: 'delete', path: ['p'], value: undefined, oldValue: 'old', object: o });
 });
@@ -48,7 +48,7 @@ suite.runTest({ name: 'verify object - root - delete' }, test => {
 suite.runTest({ name: 'verify object - deep - delete' }, test => {
 	let c;
 	const o = Observable.from({ a: { p: 'old' } });
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	delete o.a.p;
 	assertObjectEqual(test, c, { type: 'delete', path: ['a', 'p'], value: undefined, oldValue: 'old', object: o.a });
 });
@@ -58,7 +58,7 @@ suite.runTest({ name: 'verify object - deep - delete' }, test => {
 suite.runTest({ name: 'verify array - root - insert' }, test => {
 	let c;
 	const o = Observable.from([]);
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o.push('new');
 	assertObjectEqual(test, c, { type: 'insert', path: [0], value: 'new', oldValue: undefined, object: o });
 });
@@ -66,7 +66,7 @@ suite.runTest({ name: 'verify array - root - insert' }, test => {
 suite.runTest({ name: 'verify array - deep - insert' }, test => {
 	let c;
 	const o = Observable.from([[]]);
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o[0].push('new');
 	assertObjectEqual(test, c, { type: 'insert', path: [0, 0], value: 'new', oldValue: undefined, object: o[0] });
 });
@@ -74,7 +74,7 @@ suite.runTest({ name: 'verify array - deep - insert' }, test => {
 suite.runTest({ name: 'verify array - root - update' }, test => {
 	let c;
 	const o = Observable.from(['old']);
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o[0] = 'new';
 	assertObjectEqual(test, c, { type: 'update', path: ['0'], value: 'new', oldValue: 'old', object: o });
 });
@@ -82,7 +82,7 @@ suite.runTest({ name: 'verify array - root - update' }, test => {
 suite.runTest({ name: 'verify array - deep - update' }, test => {
 	let c;
 	const o = Observable.from([['old']]);
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o[0][0] = 'new';
 	assertObjectEqual(test, c, { type: 'update', path: [0, '0'], value: 'new', oldValue: 'old', object: o[0] });
 });
@@ -90,7 +90,7 @@ suite.runTest({ name: 'verify array - deep - update' }, test => {
 suite.runTest({ name: 'verify array - root - delete' }, test => {
 	let c;
 	const o = Observable.from(['old']);
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o.pop();
 	assertObjectEqual(test, c, { type: 'delete', path: [0], value: undefined, oldValue: 'old', object: o });
 });
@@ -98,7 +98,7 @@ suite.runTest({ name: 'verify array - root - delete' }, test => {
 suite.runTest({ name: 'verify array - deep - delete' }, test => {
 	let c;
 	const o = Observable.from([['old']]);
-	o.observe(cs => { c = cs[0]; })
+	Observable.observe(o, cs => { c = cs[0]; })
 	o[0].pop();
 	assertObjectEqual(test, c, { type: 'delete', path: [0, 0], value: undefined, oldValue: 'old', object: o[0] });
 });