+++ tests/unobserve.js
@@ -1,7 +1,7 @@
 import { getSuite } from '../../node_modules/just-test/dist/just-test.js';
 import { Observable } from '../../src/object-observer.js';
 
-const suite = getSuite({ name: 'Testing unobserving/removal of observed object' });
+const suite = getSuite({ name: 'Testing unobserving/removal of observed object (static API)' });
 
 suite.runTest({ name: 'test unobserve - single observer - explicit unobserve' }, () => {
 	const
@@ -11,13 +11,13 @@ suite.runTest({ name: 'test unobserve - single observer - explicit unobserve' },
 		};
 	let cntr = 0;
 
-	oo.observe(observer);
+	Observable.observe(oo, observer);
 
 	oo.some = 'thing';
 	if (cntr !== 1) throw new Error('preliminary check failed - observer was not invoked');
 
 	cntr = 0;
-	oo.unobserve(observer);
+	Observable.unobserve(oo, observer);
 	oo.some = 'true';
 	if (cntr > 0) throw new Error('unobserve failed, expected 0 callbacks, found ' + cntr);
 });
@@ -34,8 +34,8 @@ suite.runTest({ name: 'test unobserve - few observers - explicit unobserve' }, (
 	let cntrA = 0,
 		cntrB = 0;
 
-	oo.observe(observerA);
-	oo.observe(observerB);
+	Observable.observe(oo, observerA);
+	Observable.observe(oo, observerB);
 
 	oo.some = 'thing';
 	if (cntrA !== 1) throw new Error('preliminary check failed - observerA was not invoked');
@@ -43,14 +43,14 @@ suite.runTest({ name: 'test unobserve - few observers - explicit unobserve' }, (
 
 	cntrA = 0;
 	cntrB = 0;
-	oo.unobserve(observerA);
+	Observable.unobserve(oo, observerA);
 	oo.some = 'true';
 	if (cntrA > 0) throw new Error('unobserve failed, expected 0 callbacks for unobserved, found ' + cntrA);
 	if (cntrB !== 1) throw new Error('unobserve failed, expected 1 callback for the left alone after unobserve, found ' + cntrB);
 
 	cntrA = 0;
 	cntrB = 0;
-	oo.unobserve(observerB);
+	Observable.unobserve(oo, observerB);
 	oo.some = 'back';
 	if (cntrA > 0) throw new Error('unobserve failed, expected 0 callbacks for unobserved, found ' + cntrA);
 	if (cntrB > 0) throw new Error('unobserve failed, expected 0 callbacks for unobserved, found ' + cntrB);
@@ -68,8 +68,8 @@ suite.runTest({ name: 'test unobserve - unobserve few' }, () => {
 	let cntrA = 0,
 		cntrB = 0;
 
-	oo.observe(observerA);
-	oo.observe(observerB);
+	Observable.observe(oo, observerA);
+	Observable.observe(oo, observerB);
 
 	oo.some = 'thing';
 	if (cntrA !== 1) throw new Error('preliminary check failed - observerA was not invoked');
@@ -77,7 +77,7 @@ suite.runTest({ name: 'test unobserve - unobserve few' }, () => {
 
 	cntrA = 0;
 	cntrB = 0;
-	oo.unobserve(observerA, observerB);
+	Observable.unobserve(oo, observerA, observerB);
 	oo.some = 'true';
 	if (cntrA > 0) throw new Error('unobserve failed, expected 0 callbacks for unobserved, found ' + cntrA);
 	if (cntrB > 0) throw new Error('unobserve failed, expected 0 callbacks for unobserved, found ' + cntrB);
@@ -95,8 +95,8 @@ suite.runTest({ name: 'test unobserve - unobserve all' }, () => {
 	let cntrA = 0,
 		cntrB = 0;
 
-	oo.observe(observerA);
-	oo.observe(observerB);
+	Observable.observe(oo, observerA);
+	Observable.observe(oo, observerB);
 
 	oo.some = 'thing';
 	if (cntrA !== 1) throw new Error('preliminary check failed - observerA was not invoked');
@@ -104,7 +104,7 @@ suite.runTest({ name: 'test unobserve - unobserve all' }, () => {
 
 	cntrA = 0;
 	cntrB = 0;
-	oo.unobserve();
+	Observable.unobserve(oo);
 	oo.some = 'true';
 	if (cntrA > 0) throw new Error('unobserve failed, expected 0 callbacks for unobserved, found ' + cntrA);
 	if (cntrB > 0) throw new Error('unobserve failed, expected 0 callbacks for unobserved, found ' + cntrB);
@@ -118,15 +118,21 @@ suite.runTest({ name: 'test unobserve - observe, unobserve and observe again' },
 		};
 	let cntr = 0;
 
-	oo.observe(observer);
+	Observable.observe(oo, observer);
 	oo.some = 'thing';
 	if (cntr !== 1) throw new Error('preliminary check failed - observer was not invoked');
 
-	oo.unobserve();
+	Observable.unobserve(oo);
 	oo.some = 'true';
 	if (cntr !== 1) throw new Error('unobserve failed, expected callbacks for unobserved to remain 1, found ' + cntr);
 
-	oo.observe(observer);
+	Observable.observe(oo, observer);
 	oo.some = 'again';
 	if (cntr !== 2) throw new Error('preliminary check failed - observer was not invoked being added anew');
+});
+
+suite.runTest({ name: 'test unobserve - on observers set case' }, () => {
+	const oo = Observable.from({ some: 'text' });
+
+	Observable.unobserve(oo, () => { });
 });