+++ tests/workers/perf-sync-test-a.js
@@ -49,8 +49,8 @@ export default setup => {
 	}
 
 	//	add listeners/callbacks
-	po.observe(changes => changesCountA += changes.length);
-	po.observe(changes => changesCountB += changes.length);
+	Observable.observe(po, changes => changesCountA += changes.length);
+	Observable.observe(po, changes => changesCountB += changes.length);
 
 	//	mutation of existing property
 	changesCountA = 0;